refactor(wallet-modal): simplify state management and improve modularity by utilizing custom hooks
